requirements.txt
Flask==1.1.2
flask_pymongo
Flask-SQLAlchemy==2.4.1
PyJWT==1.7.1
SQLAlchemy==1.3.16
Werkzeug
datetime
flask-cors
openpyxl
openpyxl-templates
psycopg2
random_object_id
flask_expects_json
python-dateutil
